pyproject.toml
[tool.poetry]
name = "pythoneda-application-git-repositories"
version = "0.0.1a4"
description = "Application layer of PythonEDA Git Repositories package"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
packages = [{include = "PythonEDAApplicationGitRepositories"}]

[tool.poetry.dependencies]
python = "^3.7"
pythoneda-application-base = "^0.0.1a5"
pythoneda-git-repositories = "^0.0.1a4"
pythoneda-infrastructure-git-repositories = "^0.0.1a6"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
